refacator : 출력양식 변경 및 불필요한 클래스 삭제

+++ src/main/java/chess/domain/Status.java
@@ -21,11 +21,11 @@ public class Status {
 		return Team.NONE;
 	}
 
-	public int getWhiteScore() {
+	public double getWhiteScore() {
 		return white.getScore();
 	}
 
-	public int getBlackScore() {
+	public double getBlackScore() {
 		return black.getScore();
 	}
 }
